% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateSyntheticData.R
\name{generateSyntheticData}
\alias{generateSyntheticData}
\title{Generate synthetic BED data, M and GC files.}
\usage{
generateSyntheticData(dir, nchr = 5, K = 5, I = 100, J = 3)
}
\arguments{
\item{dir}{The directory for the generated datasets. If this directory already exists, it will be removed before being recreated.}

\item{nchr}{The number of chromosomes. Default: 5.}

\item{K}{The number of different transcription factors. Default: 5.}

\item{I}{The number of loci. Default: 100.}

\item{J}{The number of clusters among the loci. Default: 3.}
}
\value{
A RangedData object.
}
\description{
This function generates three set of files. ChIP BED files are stored in 'chip/', input BED files are stored in 'input/', mappability and GC files are stored in 'mgc/'. All BED files follow the name convention for the ENCODE consortium.
Each chromosome has a size of 10K, and each locus has 20 bp.
ChIP data for two celllines, each with K transcription factors, are generated. All ChIP experiments with the same celltype have the same matching input. Each ChIP experiment has randomly 1-3 replicates. Each input has 3 replicates.
}
\examples{
\dontrun{generateSyntheticData("tmpData/", nchr = 5, K = 5, I = 100, J = 3) }
}
\author{
Chandler Zuo \email{zuo@stat.wisc.edu}
}

