% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MBASIC_binary.R
\name{MBASIC.binary.all}
\alias{MBASIC.binary.all}
\alias{MBASIC.binary.fitall}
\title{Bayesian clustering model for binary state matrix with prior estimated background means.}
\usage{
MBASIC.binary.fitall(Y, Mu0, fac, allJ = NULL, allStruct = NULL,
  zeta = 0.2, maxitr = 1000, burnin = 100, family = "negbin",
  tol = 1e-04, nsig, min.count = 1, parent.id, outdir, ncores = 1)
}
\arguments{
\item{Y}{An N by I matrix containing the data from N experiments across I observation units (loci).}

\item{Mu0}{An N by I matrix for the prior estimated mean for the background state, for N experiments across the I observation units (loci).}

\item{fac}{A vector of length N denoting the experimental condition for each replicate.}

\item{allJ}{A list for the numbers of clusters for each candidate model.}

\item{allStruct}{A list of matrices indicating the levels of the signal matrix.}

\item{zeta}{The initial value of the proportion of unclustered units. Default: 0.2.}

\item{maxitr}{The maximum number of iterations in the E-M algorithm. Default: 100.}

\item{burnin}{An integer value for the number of iterations in initialization. Default: 20.}

\item{family}{The distribution of family to be used. Either "lognormal" or "negbin". See details for more information.}

\item{tol}{Tolerance for error in checking the E-M algorithm's convergence. Default: 1e-04.}

\item{nsig}{The number of mixture components for the distribution of the signal state.}

\item{parent.id}{A vector for the identifier of the parent model for each candidate model. The parent model has a more generalized structure compared to the child model, therefore, its likelihood should be smaller.}

\item{outdir}{The file directory for writing the intermediate results every 10 E-M iterations. This can be useful when the running time until final convergence is long. Default: NULL (no intermediate result is saved).}

\item{ncores}{The number of parallel sessions used.}
}
\value{
A 'MBASICFit' class object.
}
\description{
Bayesian clustering model for binary state matrix with prior estimated background means.
}
\details{
TODO
}
\examples{
\dontrun{
}
}
\author{
Chandler Zuo \email{zuo@stat.wisc.edu}
}

