% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MBASIC.R
\name{MBASIC.full}
\alias{MBASIC.full}
\title{Simultaneously fitting MBASIC model for different numbers of clusters.}
\usage{
MBASIC.full(J = NULL, struct = NULL, out = NULL, ncores = 1, ...)
}
\arguments{
\item{J}{A vector of the numbers of clusters to be identified.}

\item{struct}{A list of K by J matrix indicating the structures of each cluster. Default: NULL.}

\item{out}{The prefix of the file names to write model output information. Default: NULL, when the information is written to the R output.}

\item{ncores}{The number of CPUs to be used in parallele. Default: 1.}

\item{...}{Other parameters that are passed to \code{\link{MBASIC}}.}
}
\value{
A list object including the following fields:
\tabular{ll}{
allFits \tab A list of \linkS4class{MBASICFit} objects with different numbers of clusters and/or structures.\cr
BestFit \tab The best model with minimum BIC.\cr
Time \tab The time used for the model fitting.\cr}
}
\description{
Simultaneously fitting MBASIC model for different numbers of clusters.
}
\details{
This function fits MBASIC models with different numbers of clusters and/or different structures in parallel.
}
\examples{
## Simulate a dataset
dat.sim <- MBASIC.sim(xi = 2, family = "lognormal", I = 1000, fac = rep(1:10, each = 2), J = 3, S = 3, zeta = 0.1)
## Fit the model
dat.sim.fit <- MBASIC.full(Y = dat.sim$Y, S = 3, fac = rep(1:10, each = 2), J = 3:5, maxitr = 3, para = NULL, family = "lognormal", method = "MBASIC", zeta = 0.1, tol = 1e-6, tol.par = 0.001)
}
\author{
Chandler Zuo \email{zuo@stat.wisc.edu}
}
\seealso{
\code{\link{MBASIC}}
}

