% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MBASIC}
\alias{MBASIC}
\title{Bayesian clustering model for a state-space matrix.}
\usage{
MBASIC(Y, S, fac, J = NULL, maxitr = 100, struct = NULL, para = NULL,
  family = "lognormal", method = "em", zeta = 0.1, tol = 1e-04,
  out = NULL)
}
\arguments{
\item{Y}{An N by I matrix containing the data from N experiments across I observation units.}

\item{S}{An integer for the number of states.}

\item{fac}{A vector of levels representing the conditions of each replicate.}

\item{J}{The number of clusters to be identified.}

\item{maxitr}{The maximum number of iterations in the E-M algorithm. Default: 100.}

\item{struct}{A K by J matrix indicating the structures of each cluster.}

\item{para}{A list object that contains the true model parameters. Default: NULL. See details for more information.}

\item{family}{The distribution of family to be used. Either "lognormal" or "negbin". See details for more information.}

\item{method}{A string for the fitting method, 'naive', 'em' (default) or '2em'. See details for more information.}

\item{zeta}{The initial value for the proportion of units that are not clustered. Default: 0.1.}

\item{tol}{Tolerance for error in checking the E-M algorithm's convergence. Default: 1e-04.}

\item{out}{The file directory for writing fitting information in each E-M iteration. Default: NULL ( no information is outputted ).}
}
\value{
An object of class 'MBASICFit'.
}
\description{
This function is designed to analyze general state-space models. The data consists of observations over I units under N experiments with K different conditions. There are S states for each experiment and unit.
}
\details{
Function MBASIC currently supports two different distributional families: log-normal and negative binomial. This should be specified by the 'family' argument.\cr
For the log-normal distributions, log(Y+1) is modeled as normal distributions. For experiment n, if locus i has state s, distribution for log(Y[n,i]+1) is N( Mu[n,s], Sigma[n,s] ).\cr
For the negative binomial distributions, the meanings of Mu and Sigma are different. For experiment n, if locus i has state s, distribution of Y[n,i] is NB( Mu[n,s], Sigma[n,s] ). In this package, NB( mu, a ) denotes the negative-binomial distribution with mean mu and size a (i.e. the variance is mu*(1+mu/a) ).\cr
 The 'method' argument determines what fitting method will be used. The default is 'em', where the states and the clustering are simultaneously estimated. 'naive' and '2em' methods use 2-step algorithms. In the first step, both estimate the states for each unit by an E-M algorithm for each experiment. In the second step, 'naive' uses hierarchical clustering to cluster the units, while '2em' uses function 'MBASIC.state' to identify clusters.\cr
The 'para' argument takes a list object that is supposed to include the following fields:
\tabular{ll}{
W \tab A K by (J*S) matrix. The (k,J*(s-1)+j)-th entry is the probability that the units in cluster j has state s in the k-th experiment.\cr
Z \tab An I by J matrix. The (i,j)-th entry is the indicator whether the i-th unit belongs to cluster j.\cr
Theta \tab A K by (I*S) matrix. The (k,I*(s-1)+i)-th entry is the probability that the i-th unit has state s in the k-th experiment.\cr
non.id \tab A binary vector of length I. The i-th entry is the indicator whether the i-th unit does not belong to any cluster.
}
This argument is intended to carry the true parameters in simulation studies. If it is not null, then the model also computes a number of metrics that describes the error in model fitting. Users should be cautious that the order of the rows and columns of matrices in the fields of para should match the Y matrix.
}
\examples{
## Simulate a dataset
dat.sim <- MBASIC.sim( xi = 2, family = "lognormal", I = 1000, fac = rep( 1:10, each = 2 ), J = 3, S = 3, zeta = 0.1 )
## Fit the model
dat.sim.fit <- MBASIC( Y = dat.sim$Y, S = 3, fac = rep( 1:10, each = 2), J = 3, maxitr = 3, para = NULL, family = "lognormal", method = "em", zeta = 0.1, tol = 1e-04)
}
\author{
Chandler Zuo \email{zuo@stat.wisc.edu}
}

