% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MBASIC_simulate.R
\name{MBASIC.sim.state}
\alias{MBASIC.sim.state}
\title{Simulate the state matrix for general MBASIC models.}
\usage{
MBASIC.sim.state(I, K, J, S = 2, struct = NULL, delta = NULL,
  delta.non = NULL, zeta = 0.1, statemap = NULL)
}
\arguments{
\item{I}{An integer for the total number of units.}

\item{K}{An integer for the number of different experimental conditions.}

\item{J}{An integer for the number of clusters.}

\item{S}{An integer for the number of states.}

\item{struct}{An K by J integer matrix. The j-th column denotes the levels for the cluster level parameter. See details for more information. Default: NULL.}

\item{delta}{A vector the same length as statemap, or NULL. This is the dirichlet prior parameter to simulate the probability across the components for each CLUSTERED unit and each experiment. If NULL, rep(0.1, length(statemap)) is used.}

\item{delta.non}{A vector of length S, or NULL. This is the dirichlet prior parameter to simulate the probability across the components for each UNCLUSTERED unit and each experiment. If NULL, rep(0.1, length(statemap)) is used.}

\item{zeta}{The probability that each unit does not belong to any cluster. Default: 0.1.}

\item{statemap}{A vector consisted of 1, 2, ..., S, representing which state each component belongs to. Default: NULL.}
}
\value{
A list containing:
\tabular{ll}{
 Theta \tab A K by I matrix. The (k,i)-th element is the indicator of whether the i-th unit is binding for condition k.\cr
 W \tab A K by J matrix. Each row is the indicators of the loci sets related an individual replicate. \cr
 Z \tab An I by J matrix. Each column is the indicator for an individual loci set.\cr
 delta \tab Same as the input argument \code{delta}.\cr
 zeta \tab Same as the input argumnent \code{zeta}.\cr
 non.id \tab A vector of length I indicating the number of each unit not to belong to any cluster.\cr
}
}
\description{
Simulate for the Theta matrix
}
\examples{
dat.sim <- MBASIC.sim.state(I = 100, K = 10, J = 3)
}
\author{
Chandler Zuo \email{zuo@stat.wisc.edu}
}

