% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/visualize.R
\docType{methods}
\name{plot}
\alias{MBASICFit-methods}
\alias{plot}
\alias{plot,}
\alias{plot,MBASICFit,missing-method}
\title{Plot the state probability for a fitted MBASIC model.}
\usage{
\S4method{plot}{MBASICFit,missing}(x, y = numeric(0), state = 1,
  slot = "Theta", xlab = NULL, ylab = NULL, main = NULL, Rowv = FALSE,
  Colv = FALSE, trace = "none", dendrogram = "none", ...)
}
\arguments{
\item{x}{An \linkS4class{MBASICFit} object.}

\item{y}{Missing.}

\item{state}{The probability for this state is plotted.}

\item{slot}{The slot to be plotted, "Theta" (default), "W", "Mu", or "Sigma".}

\item{xlab}{The x-axis label.}

\item{ylab}{The y-axis label.}

\item{main}{The title.}

\item{Rowv,Colv,trace,dendrogram,...}{Other arguments passed to \code{\link[gplots]{heatmap.2}}.}

\item{...}{Other arguments passed to the function \code{\link[gplots]{heatmap.2}}.}
}
\value{
A heatmap showing the probability for a given state.
}
\description{
Plot the state probability for a fitted MBASIC model.
}
\details{
This function provides four differerent kinds of plots.\cr
If \code{slot='Theta'}, a heatmap for the probability for each unit to have state \code{state} under each condition is drawn.\cr
If \code{slot='W'}, a heatmap for the probability for each cluster to have state \code{state} under each condition is drawn.\cr
If \code{slot=}'Mu' or 'Sigma', the range of the fitted 'Mu' or 'Sigma' parameters among different replicates for the same condition is drawn.\cr
}
\author{
Chander Zuo \email{zuo@stat.wisc.edu}
}

