% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utility.R
\name{write.out}
\alias{write.out}
\title{Write the message to an output file.}
\usage{
write.out(out, msg)
}
\arguments{
\item{out}{Either the file name, or NULL. If NULL, the message is written to the standard output.}

\item{msg}{The string for the message to be written.}
}
\description{
Write the message to an output buffer.
}
\author{
Chandler Zuo \email{zuo@stat.wisc.edu}
}

